fix: copy the declarations to the dist
Was broken after the migration
